tests/requirements-test.txt
# Tests require pyyaml and pytest
pytest>=8.0
pyyaml>=6.0
